Fixed float64 issue in cython correlograms. Adding channel group parameter when loading view data. Got rid of trace view temporarily.

+++ klustaviewa/stats/correlograms_cython.pyx
@@ -1,7 +1,7 @@
 import numpy as np
 cimport numpy as np
-DTYPE = np.float32
-ctypedef np.float32_t DTYPE_t
+DTYPE = np.float64
+ctypedef np.float64_t DTYPE_t
 DTYPEI = np.int32
 ctypedef np.int32_t DTYPEI_t
